backend/requirements.txt
annotated-doc==0.0.3
annotated-types==0.7.0
anyio==4.11.0
certifi==2025.10.5
charset-normalizer==3.4.4
click==8.3.0
distro==1.9.0
fastapi==0.120.3
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
jiter==0.11.1
openai==2.6.1
psycopg2-binary==2.9.11
pydantic==2.12.3
pydantic_core==2.41.4
python-dotenv==1.2.1
requests==2.32.5
sniffio==1.3.1
SQLAlchemy==2.0.44
starlette==0.49.1
tqdm==4.67.1
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.5.0
uvicorn==0.38.0
beautifulsoup4==4.12.3
pgvector==0.3.6
pytest==9.0.2
requests==2.32.5
pytest-playwright
uvicorn
